requirements.txt
pyseagull
Pillow